# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "AnonymousSessionResponse": {
        "additionalProperties": false,
        "properties": {
          "authenticated": {
            "enum": [
              false
            ],
            "type": "boolean"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "authenticated",
          "permissions"
        ],
        "type": "object"
      },
...
      "ApiKeyIdentity": {
        "properties": {
          "id": {
            "format": "int64",
            "minimum": 1,
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "type": "array",
            "uniqueItems": true
          }
        },
        "required": [
          "id",
          "name",
          "scopes"
        ],
        "type": "object"
      },
...
      "BearerSessionResponse": {
        "additionalProperties": false,
        "properties": {
          "api_key": {
            "$ref": "#/components/schemas/ApiKeyIdentity"
          },
          "authenticated": {
            "enum": [
              true
            ],
            "type": "boolean"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "authenticated",
          "api_key",
          "permissions"
        ],
        "type": "object"
      },
...
      "BrowserSessionResponse": {
        "additionalProperties": false,
        "properties": {
          "authenticated": {
            "enum": [
              true
            ],
            "type": "boolean"
          },
          "csrf_token": {
            "type": "string"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "user": {
            "$ref": "#/components/schemas/UserResource"
          }
        },
        "required": [
          "authenticated",
          "user",
          "csrf_token",
          "permissions"
        ],
        "type": "object"
      },
...
      "LoginInput": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "type": "string"
          },
          "remember": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "type": "object"
      },
...
      "PasswordInput": {
        "additionalProperties": false,
        "properties": {
          "current_password": {
            "type": "string"
          },
          "new_password": {
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "type": "string"
          },
          "status": {
            "format": "int32",
            "maximum": 599,
            "minimum": 100,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "format": "uri-reference",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "type": "object"
      },
...
      "SessionCreatedResponse": {
        "properties": {
          "csrf_token": {
            "type": "string"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "user": {
            "$ref": "#/components/schemas/UserResource"
          }
        },
        "required": [
          "user",
          "csrf_token",
          "permissions"
        ],
        "type": "object"
      },
      "SessionResponse": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/BrowserSessionResponse"
          },
          {
            "$ref": "#/components/schemas/BearerSessionResponse"
          },
          {
            "$ref": "#/components/schemas/AnonymousSessionResponse"
          }
        ]
      },
...
      "UserResource": {
        "properties": {
          "id": {
            "format": "int64",
            "minimum": 1,
            "type": "integer"
          },
          "password_change_required": {
            "type": "boolean"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "role",
          "password_change_required"
        ],
        "type": "object"
      },
      "UserRole": {
        "enum": [
          "user",
          "admin",
          "owner"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "in": "cookie",
        "name": "racebin_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Racebin contributors"
    },
    "description": "API-first, self-hosted paste bin and file share.",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "Racebin API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/account/password": {
      "patch": {
        "operationId": "change_password",
        "parameters": [
          {
            "description": "Current browser session CSRF token",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Password changed and existing sessions revoked"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid password or browser session required"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Current password is incorrect"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "CSRF validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "account"
        ]
      }
    },
...
    "/session": {
      "delete": {
        "operationId": "logout",
        "parameters": [
          {
            "description": "Current browser session CSRF token",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Session deleted",
            "headers": {
              "Set-Cookie": {
                "description": "Expired Racebin session cookie",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Browser session required"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "CSRF validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "account"
        ]
      },
      "get": {
        "operationId": "get_session",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Current browser session, bearer-key identity, or anonymous state"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid bearer credential"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {},
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "account"
        ]
      },
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreatedResponse"
                }
              }
            },
            "description": "Browser session created",
            "headers": {
              "Set-Cookie": {
                "description": "HTTP-only Racebin session cookie",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Malformed credentials"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid credentials"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Login rate limit exceeded",
            "headers": {
              "Retry-After": {
                "description": "Seconds before another attempt",
                "schema": {
                  "format": "int64",
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {}
        ],
        "tags": [
          "account"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
}